Add go-homedir instead of os/user
This should fix #225 but in full disclosure, I don't have a win
system handy at the moment. That said, no regressions emerge when
running ```go test ./...``` so I'm kinda counting on coverage to
show signs. I can probably test on a win box in the next couple of
days, but welcome others doing so.

Would love :eyes: on this from @tj to make sure the actual issue
was addressed.

Just trying to contribute where I can and this seemed easy...